rollup: move code to plutus-emulator
