+++ plutus-emulator/plutus-emulator.cabal
@@ -45,6 +45,9 @@ library
         Wallet.Emulator
         Wallet.Emulator.Types
         Wallet.Emulator.Generators
+        Wallet.Rollup
+        Wallet.Rollup.Types
+        Wallet.Rollup.Render
     hs-source-dirs: src
     default-language: Haskell2010
     build-depends:
@@ -67,6 +70,7 @@ library
         mtl -any,
         natural-transformation -any,
         operational -any,
+        prettyprinter -any,
         serialise -any,
         servant -any,
         servant-client -any,
