+++ plutus-playground-lib/plutus-playground-lib.cabal
@@ -28,8 +28,6 @@ library
   exposed-modules:
       Playground.Contract
       Playground.API
-      Playground.Rollup
-      Playground.Rollup.Render
       Playground.Types
       Playground.TH
       Playground.Interpreter.Util
