+++ plutus-playground-lib/src/Playground/Types.hs
@@ -11,24 +11,22 @@
 module Playground.Types where
 
 import           Control.Lens                 (makeLenses)
-import           Data.Aeson                   (FromJSON, FromJSONKey, ToJSON, ToJSONKey)
+import           Data.Aeson                   (FromJSON, ToJSON)
 import qualified Data.Aeson                   as JSON
 import           Data.List.NonEmpty           (NonEmpty ((:|)))
-import           Data.Map                     (Map)
 import           Data.Text                    (Text)
 import           GHC.Generics                 (Generic)
 import           Language.Haskell.Interpreter (CompilationError, SourceCode)
 import qualified Language.Haskell.Interpreter as HI
 import qualified Language.Haskell.TH.Syntax   as TH
-import           Ledger                       (Address, Blockchain, PubKey, Tx, TxId, TxIn, TxOut,
-                                               TxOutOf (TxOutOf, txOutAddress, txOutType),
-                                               TxOutType (PayToPubKey, PayToScript), fromSymbol)
+import           Ledger                       (Blockchain, PubKey, Tx, TxId, fromSymbol)
 import qualified Ledger.Ada                   as Ada
 import           Ledger.Scripts               (ValidatorHash)
 import           Ledger.Value                 (TokenName)
 import qualified Ledger.Value                 as V
 import           Schema                       (FormSchema)
 import           Wallet.Emulator.Types        (EmulatorEvent, Wallet, walletPubKey)
+import           Wallet.Rollup.Types          (AnnotatedTx)
 
 data KnownCurrency =
     KnownCurrency
@@ -122,48 +120,4 @@ data PlaygroundError
     deriving (Show, Generic)
     deriving anyclass (ToJSON, FromJSON)
 
-------------------------------------------------------------
-data SequenceId =
-    SequenceId
-        { slotIndex :: Int
-        , txIndex   :: Int
-        }
-    deriving (Eq, Ord, Show, Generic)
-    deriving anyclass (FromJSON, ToJSON)
-
-data DereferencedInput =
-    DereferencedInput
-        { originalInput :: TxIn
-        , refersTo      :: TxOut
-        }
-    deriving (Eq, Show, Generic)
-    deriving anyclass (FromJSON, ToJSON)
-
-data BeneficialOwner
-    = OwnedByPubKey PubKey
-    | OwnedByScript Address
-    deriving (Eq, Show, Ord, Generic)
-    deriving anyclass (FromJSON, ToJSON, FromJSONKey, ToJSONKey)
-
-toBeneficialOwner :: TxOut -> BeneficialOwner
-toBeneficialOwner TxOutOf {txOutType, txOutAddress} =
-    case txOutType of
-        PayToPubKey pubKey -> OwnedByPubKey pubKey
-        PayToScript _      -> OwnedByScript txOutAddress
-
-data AnnotatedTx =
-    AnnotatedTx
-        { sequenceId         :: SequenceId
-        , txId               :: TxId
-        , tx                 :: Tx
-        , dereferencedInputs :: [DereferencedInput]
-        , balances           :: Map BeneficialOwner V.Value
-        }
-    deriving (Eq, Show, Generic)
-    deriving anyclass (FromJSON, ToJSON)
-
 makeLenses 'EvaluationResult
-
-makeLenses 'SequenceId
-
-makeLenses 'AnnotatedTx
